Refactor MethodToolCallbackProviderTests
 - Move the existing buildToolsWithBridgeMethodReturnOnlyUserDeclaredMethods into MethodToolCallbackProviderTests
 - Delete MethodToolCallbackProviderTest
